+++ jira_clone/app/interactors/task_interactor.py
@@ -19,24 +19,35 @@ class TaskInteractor:
 
             return task
 
+        return 404
+
     def create_task(self, task_schema: TaskSchema):
         token = self.hasher.encode(task_schema)
 
-        self.repository.create_task(token)
+        if self.repository.get_task_by_token(token):
+            self.repository.create_task(token)
 
-        return token
+            return token
+
+        return 404
 
     def remove_task(self, task_schema: TaskSchema):
         token = self.hasher.encode(task_schema)
 
-        self.repository.remove_task(token)
+        if self.repository.get_task_by_token(token):
+            self.repository.remove_task(token)
 
-        return token
+            return token
+
+        return 404
 
     def update_task(self, old_task_schema: TaskSchema, new_task_schema: TaskSchema):
         old_token = self.hasher.encode(old_task_schema)
         new_token = self.hasher.encode(new_task_schema)
 
-        self.repository.update_task(old_token, new_token)
+        if self.repository.get_task_by_token(old_token):
+            self.repository.update_task(old_token, new_token)
 
-        return new_token
+            return new_token
+
+        return 404
